Moved common keras code to utils.
